requirements.txt
bios
torchvision
fc_quantization==0.0.4
